requirements.txt
# TradingView & WebSocket
tradingview-ta>=3.3.0
websocket-client>=1.6.0

# Data & Backtest
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
vectorbt>=0.26.0
backtrader>=1.9.78

# Technical indicators (live)
ta>=0.10.2

# ML (scanner)
scikit-learn>=1.3.0

# AI - OpenRouter
httpx>=0.25.0
openai>=1.0.0

# UI
PyQt6>=6.6.0

# Utils
python-dotenv>=1.0.0
schedule>=1.2.0

# Sentiment & News
feedparser>=6.0.10
textblob>=0.17.1
requests>=2.31.0

# Export
openpyxl>=3.1.0
reportlab>=4.0.0
